+++ tests/e2e/base/provider/flask_provider.py
@@ -14,49 +14,15 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-
-"""
-This module contains the Provider part of the e2e tests.
-consumer (FastAPI) -> consumer (AIOHTTP) -> provider (FastAPI + logging_with_exception)
-We also cover the usage of logging interception in this module.
-"""
-import io
 import logging
-import random
 import time
-import traceback
+import random
+import os
+from flask import Flask
 
-import uvicorn
-from fastapi import FastAPI
+from log_formatter import E2EProviderFormatter
 
-
-class E2EProviderFormatter(logging.Formatter):
-    """
-    User defined formatter should in no way be
-    interfered by the log reporter formatter.
-    """
-
-    def format(self, record):
-        result = super().format(record)
-        return f'e2e_provider:={result}'
-
-    def formatException(self, ei):  # noqa
-        """
-        Mock user defined formatter which limits the traceback depth
-        to None -> meaning it will not involve trackback at all.
-        but the agent should still be able to capture the traceback
-        at default level of 5 by ignoring the cache.
-        """
-        sio = io.StringIO()
-        tb = ei[2]
-        traceback.print_exception(ei[0], ei[1], tb, None, sio)
-        s = sio.getvalue()
-        sio.close()
-        if s[-1:] == '\n':
-            s = s[:-1]
-        return s
-
-
+app = Flask(__name__)
 formatter = E2EProviderFormatter(logging.BASIC_FORMAT)
 stream_handler = logging.StreamHandler()
 stream_handler.setFormatter(formatter)
@@ -65,12 +31,21 @@ e2e_provider_logger = logging.getLogger('__name__')
 e2e_provider_logger.setLevel(logging.INFO)
 e2e_provider_logger.addHandler(stream_handler)
 
-app = FastAPI()
+
+@app.route('/pid', methods=['GET'])
+def pid():
+    """
+    This endpoint is used to get the pid of the provider, for e2e tests.
+    Returns: The instance name of the provider.
+    """
+    from skywalking import config
+    # Instance name is dynamically modified by uwsgi hook/os.fork handler
+
+    return {'instance': config.agent_instance_name}
 
 
-@app.get('/artist')
-@app.post('/artist')
-async def application():
+@app.route('/artist-provider', methods=['POST'])
+def artist():
     time.sleep(random.random())
     # Exception is reported with trackback depth of 5 (default)
     try:
@@ -88,9 +63,12 @@ async def application():
     # Debug is not reported according to default agent setting
     e2e_provider_logger.debug('E2E Provider Debug, this is not reported!')
 
-    return {'artist': 'Luis Fonsi'}
+    return {
+        'artist': 'Luis Fonsi',
+        'pid': os.getpid()
+    }
 
 
 if __name__ == '__main__':
     # noinspection PyTypeChecker
-    uvicorn.run(app, host='0.0.0.0', port=9090)
+    app.run(host='0.0.0.0', port=9090)