+++ CHANGELOG.md
@@ -8,7 +8,8 @@
   please check with the latest official documentation before upgrading. (#273, #282)
   https://skywalking.apache.org/docs/skywalking-python/v1.0.0/en/setup/configuration/
   - **BREAKING**: All agent core capabilities are now covered by test cases and enabled by default (Trace, Log, PVM runtime metrics, Profiler)
-
+  - **BREAKING**: DockerHub Python agent images since v1.0.0 will no longer include the `run` part in `ENTRYPOINT ["sw-python", "run"]`, 
+  user should prefix their command with `[-d/--debug] run [-p/--prefork] <Command>` for extra flexibility.
 
 - Feature:
   - Add support for Python 3.11 (#285)
@@ -19,7 +20,8 @@
   - Add support for the tags of Virtual Cache for Redis (#263)
   - Add a new configuration `kafka_namespace` to prefix the kafka topic names (#277)
   - Add log reporter support for loguru (#276)
-  - Add **experimental** support for explicit os.fork(), restarts agent in new process (#286)
+  - Add **experimental** support for explicit os.fork(), restarts agent in forked process (#286)
+  - Add **experimental** sw-python CLI `sw-python run [-p]` flag (-p/--prefork) to enable non-intrusive uWSGI and Gunicorn postfork support (#288)
 
 - Plugins:
   - Add aioredis, aiormq, amqp, asyncpg, aio-pika, kombu RMQ plugins (#230 Missing test coverage)